Move connect into its own file
